+++ components/tidb_query/src/expr_util/mod.rs
@@ -5,4 +5,3 @@
 pub mod conv;
 pub mod like;
 pub mod miscellaneous;
-pub mod time;
